update forget password and reset password ui validation
