+++ lib/main.dart
@@ -1,10 +1,10 @@
 import 'package:bookazon/resources/localization/generated/l10n.dart';
 import 'package:bookazon/resources/router/app_router.dart';
-import 'package:bookazon/view/screens/email_verification_page.dart';
-import 'package:bookazon/view/screens/forgetpassword_page.dart';
+import 'package:bookazon/view/screens/email%20verification/email_verification_page.dart';
+import 'package:bookazon/view/screens/forget%20password/forgetpassword_page.dart';
 import 'package:bookazon/view/screens/login/login_page.dart';
-import 'package:bookazon/view/screens/reset_password_page.dart';
-import 'package:bookazon/view/screens/signup_page.dart';
+import 'package:bookazon/view/screens/reset%20password/reset_password_page.dart';
+import 'package:bookazon/view/screens/signup/signup_page.dart';
 import 'package:bookazon/view/screens/splash/splash_page.dart';
 import 'package:bookazon/view_model/auth/auth_cubit.dart';
 import 'package:bookazon/view_model/onboarding/onboarding_cubit.dart';
@@ -53,7 +53,7 @@ class MyApp extends StatelessWidget {
               useMaterial3: true,
             ),
             onGenerateRoute: RouteGenerate.getRoute,
-            home: ResetPasswordPage(),//const SplashPage(),
+            home: const SplashPage(),
           ),
         );
       },
